requirements.txt
customtkinter
pyautogui
keyboard
Pillow
